requirements.txt
python-telegram-bot
discord.py
apscheduler
python-dotenv
tzdata